Change the more info buttons
This commit reworks how the info buttons work. I made it so that the
buttons are all visible, but change in size if they are selected. This
will make the website more clean, while still implementing the JS file.
Styling was also implemented for smaller screens. The outline of
everything isn't properly outlining.